refactored custom actions and changed overview msg
