--- src/rasa/config.yml
@@ -1,32 +1,0 @@
-# Configuration for Rasa NLU.
-# https://rasa.com/docs/rasa/nlu/components/
-language: de
-pipeline:
-  - name: WhitespaceTokenizer
-  - name: RegexFeaturizer
-  - name: LexicalSyntacticFeaturizer
-  - name: CountVectorsFeaturizer
-  - name: CountVectorsFeaturizer
-    analyzer: "char_wb"
-    min_ngram: 1
-    max_ngram: 4
-  - name: DIETClassifier
-    epochs: 100
-  - name: EntitySynonymMapper
-  - name: ResponseSelector
-    epochs: 100
-  - name: FallbackClassifier
-    threshold: 0.7
-
-# Configuration for Rasa Core.
-# https://rasa.com/docs/rasa/core/policies/
-policies:
-  - name: MemoizationPolicy
-    max_history: 5
-  - name: TEDPolicy
-    max_history: 5
-    epochs: 100
-  - name: RulePolicy
-    core_fallback_threshold: 0.7
-    core_fallback_action_name: "action_default_fallback"
-    enable_fallback_prediction: True
