--- src/rasa/endpoints.yml
@@ -1,37 +1,0 @@
-# This file contains the different endpoints your bot can use.
-
-# Server where the models are pulled from.
-# https://rasa.com/docs/rasa/user-guide/configuring-http-api/#fetching-models-from-a-server/
-
-#models:
-#  url: http://my-server.com/models/default_core@latest
-#  wait_time_between_pulls:  10   # [optional](default: 100)
-
-# Server which runs your custom actions.
-# https://rasa.com/docs/rasa/core/actions/#custom-actions/
-
-action_endpoint:
-  url: "http://actions:5055/webhook"
-  # url: "http://localhost:5055/webhook"
-
-# Tracker store which is used to store the conversations.
-# By default the conversations are stored in memory.
-# https://rasa.com/docs/rasa/api/tracker-stores/
-
-tracker_store:
-   type: SQL
-   dialect: "postgresql"
-   url: ${AYA_TRACKER_DB_HOST}
-   port: ${AYA_TRACKER_DB_PORT}
-   db: ${AYA_TRACKER_DB_DB}
-   username: ${AYA_TRACKER_DB_USER}
-   password: ${AYA_TRACKER_DB_PW}
-
-# Event broker which all conversation events should be streamed to.
-# https://rasa.com/docs/rasa/api/event-brokers/
-
-#event_broker:
-#  url: localhost
-#  username: username
-#  password: password
-#  queue: queue
