--- src/rasa/results/failed_stories.md
@@ -1,1 +1,0 @@
-<!-- All stories passed -->
