--- src/rasa/results/intent_errors.json
@@ -1,10 +1,0 @@
-[
-  {
-    "text": "/start",
-    "intent": "start",
-    "intent_prediction": {
-      "name": "uebersicht",
-      "confidence": 0.3590986430644989
-    }
-  }
-]
