--- src/rasa/rules.yml
@@ -1,5 +1,0 @@
-rules:
-- rule: Utter default message when user message has low NLU confidence
-  steps:
-  - intent: nlu_fallback
-  - action: utter_default
