+++ src/rasa/src/credentials.yml
@@ -1,0 +1,35 @@
+# This file contains the credentials for the voice & chat platforms
+# which your bot is using.
+# https://rasa.com/docs/rasa/user-guide/messaging-and-voice-channels/
+
+# rest:
+# #  # you don't need to provide anything here - this channel doesn't
+# #  # require any credentials
+
+callback:
+  # URL to which Core will send the bot responses
+  url: "http://flask:5000/bot"
+
+#facebook:
+#  verify: "<verify>"
+#  secret: "<your secret>"
+#  page-access-token: "<your page access token>"
+
+#slack:
+#  slack_token: "<your slack token>"
+#  slack_channel: "<the slack channel>"
+
+# socketio:
+#  user_message_evt: user_uttered
+#  bot_message_evt: bot_uttered
+#  session_persistence: true
+
+#mattermost:
+#  url: "https://<mattermost instance>/api/v4"
+#  token: "<bot token>"
+#  webhook_url: "<callback URL>"
+
+# This entry is needed if you are using Rasa X. The entry represents credentials 
+# for the Rasa X "channel", i.e. Talk to your bot and Share with guest testers.
+rasa:
+  url: "http://localhost:5002/api"
