+++ src/rasa/src/data/stories.yml
@@ -1,0 +1,140 @@
+version: "2.0"
+stories:
+- story: start
+  steps:
+  - intent: start
+  - action: utter_start
+  - intent: mood_great
+  - action: utter_happy
+- story: how are you
+  steps:
+  - intent: greet
+  - action: utter_greet
+  - intent: how_are_you
+  - action: utter_I_am_fine
+  - intent: affirm
+  - action: utter_ok
+- story: happy path
+  steps:
+  - intent: greet
+  - action: utter_greet
+  - intent: mood_great
+  - action: utter_happy
+- story: sad path
+  steps:
+  - intent: greet
+  - action: utter_greet
+  - intent: mood_unhappy
+  - action: utter_express_empathy
+- story: say no problem
+  steps:
+  - intent: no_problem
+  - action: utter_no_problem
+- story: say goodbye
+  steps:
+  - intent: goodbye
+  - action: utter_goodbye
+- story: fasten start yes
+  steps:
+  - intent: fasten
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 0
+  - action: utter_ask_fast
+  - intent: affirm
+  - action: utter_start_fast
+  - action: action_start_fast
+  - slot_was_set:
+    - is_fasting: 1
+- story: fasten start no
+  steps:
+  - intent: fasten
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 0
+  - action: utter_ask_fast
+  - intent: deny
+  - action: utter_ok
+- story: fasten start while fasting
+  steps:
+  - intent: fasten
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 1
+  - action: utter_ask_end_fast
+- story: fast end without fasting yes
+  steps:
+  - intent: fast_end
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 0
+  - action: utter_info_not_fasting_ask_fast
+  - intent: affirm
+  - action: action_start_fast
+  - slot_was_set:
+    - is_fasting: 1
+  - action: utter_start_fast
+- story: fast end without fasting no
+  steps:
+  - intent: fast_end
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 0
+  - action: utter_info_not_fasting_ask_fast
+  - intent: deny
+  - action: utter_ok
+- story: fast end 1
+  steps:
+  - intent: fast_end
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 1
+  - action: utter_ask_end_fast
+  - intent: affirm
+  - action: utter_end_fast
+  - action: action_end_fast
+  - slot_was_set:
+    - is_fasting: 0
+- story: fast end 2
+  steps:
+  - intent: fast_end
+  - action: action_fasting_since
+  - slot_was_set:
+    - is_fasting: 1
+  - action: utter_ask_end_fast
+  - intent: deny
+  - action: utter_ok
+- story: fast reminder
+  steps:
+  - intent: fast_reminder
+  - action: action_fasting_since
+  - action: utter_reminder_fast
+  - action: action_set_reminder_fast
+- story: uebersicht
+  steps:
+  - intent: uebersicht
+  - action: action_send_overview
+- story: weight add yes
+  steps:
+  - intent: weight_add
+  - action: utter_ask_weight
+  - intent: affirm
+  - action: utter_ask_weight_value
+  - intent: weight_value
+  - action: action_entity_extract
+  - action: utter_noted
+- story: weight add no
+  steps:
+  - intent: weight_add
+  - action: utter_ask_weight
+  - intent: deny
+  - action: utter_ok
+- story: weight value
+  steps:
+  - intent: weight_value
+  - action: action_entity_extract
+  - action: utter_noted
+- story: rezepte
+  steps:
+  - intent: rezepte
+  - action: utter_send_recipes
