+++ src/rasa/src/results/failed_stories.md
@@ -1,0 +1,1 @@
+<!-- All stories passed -->
