--- src/rasa/story_graph.dot
@@ -1,72 +1,0 @@
-digraph  {
-0 [class="start active", fillcolor=green, fontsize=12, label=START, style=filled];
-1 [class="", fontsize=12, label=action_fasting_since];
-7 [class="", fontsize=12, label=action_send_fasting_log];
-"-9" [class=ellipsis, label="..."];
-8 [class="", fontsize=12, label=utter_ask_weight];
-"-11" [class=ellipsis, label="..."];
-10 [class="", fontsize=12, label=action_entity_extract];
-"-12" [class=ellipsis, label="..."];
-11 [class="", fontsize=12, label=utter_start];
-"-13" [class=ellipsis, label="..."];
-12 [class="", fontsize=12, label=utter_send_recipes];
-"-14" [class=ellipsis, label="..."];
-13 [class=active, fontsize=12, label=utter_greet];
-"-17" [class=ellipsis, label="..."];
-16 [class="", fontsize=12, label=utter_no_problem];
-"-18" [class=ellipsis, label="..."];
-17 [class="", fontsize=12, label=utter_goodbye];
-"-19" [class=ellipsis, label="..."];
-"-22" [class=ellipsis, label="..."];
-22 [class=active, fontsize=12, label=action_fasting_since];
-23 [class="dashed active", fontsize=12, label=action_default_fallback];
-24 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-25 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-26 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-27 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-28 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-29 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-30 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-31 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-32 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-33 [class="intent active", fillcolor=lightblue, label=hi, shape=rect, style=filled];
-34 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-35 [class=intent, fillcolor=lightblue, label=None, shape=rect, style=filled];
-36 [class="intent active", fillcolor=lightblue, label=start, shape=rect, style=filled];
-0 -> 24  [class="", key=0];
-0 -> 25  [class="", key=0];
-0 -> 26  [class="", key=0];
-0 -> 27  [class="", key=0];
-0 -> 28  [class="", key=0];
-0 -> 29  [class="", key=0];
-0 -> 30  [class="", key=0];
-0 -> 31  [class="", key=0];
-0 -> 32  [class="", key=0];
-0 -> 33  [class=active, key=0];
-0 -> 34  [class="", key=0];
-0 -> 35  [class="", key=0];
-1 -> "-22"  [class="", key=NONE, label=""];
-7 -> "-9"  [class="", key=NONE, label=""];
-8 -> "-11"  [class="", key=NONE, label=""];
-10 -> "-12"  [class="", key=NONE, label=""];
-11 -> "-13"  [class="", key=NONE, label=""];
-12 -> "-14"  [class="", key=NONE, label=""];
-13 -> "-17"  [class="", key=NONE, label=""];
-13 -> 36  [class=active, key=0];
-16 -> "-18"  [class="", key=NONE, label=""];
-17 -> "-19"  [class="", key=NONE, label=""];
-22 -> 23  [class=active, key=NONE, label=""];
-24 -> 1  [class="", key=0];
-25 -> 1  [class="", key=0];
-26 -> 1  [class="", key=0];
-27 -> 1  [class="", key=0];
-28 -> 7  [class="", key=0];
-29 -> 8  [class="", key=0];
-30 -> 10  [class="", key=0];
-31 -> 11  [class="", key=0];
-32 -> 12  [class="", key=0];
-33 -> 13  [class=active, key=0];
-34 -> 16  [class="", key=0];
-35 -> 17  [class="", key=0];
-36 -> 22  [class=active, key=0];
-}
