+++ src/requirements_rasa_sdk.txt
@@ -1,0 +1,2 @@
+# for base functions
+numpy==1.18.5
